Configurable sample rate: 2.048M / 1.024M from the settings menu
The server accepts exactly one SetSampleRate per connection (measured:
a first-thing 1.024M request streams at 1.026M; mid-stream or repeated
switches destabilize it to 1.6-2.2M). The menu cycles between 2.048M
(±128 kHz span) and 1.024M (±64 kHz, NFM-friendly); switching
re-dimensions the whole DSP chain (rates are now variables set via
dsp.SetIQRate: IF ÷8, audio ÷4, resampler ratio follows), reconnects,
and sends the rate as the connection's first command. Persisted in
sdrg35xx.ini as rate=.

+++ internal/dsp/demo.go
@@ -30,17 +30,17 @@ func RunDemo(ctx context.Context, getChain func() *Chain, sink func([]float32)) 
 		for i := 0; i < blockSamples; i++ {
 			t++
 			// Wanted FM station: 1 kHz tone, deviation per mode.
-			m := math.Sin(2 * math.Pi * 1000 * float64(t) / IQRate)
-			phaseMain += 2 * math.Pi * dev * 0.6 * m / IQRate
+			m := math.Sin(2 * math.Pi * 1000 * float64(t) / float64(float64(IQRate)))
+			phaseMain += 2 * math.Pi * dev * 0.6 * m / float64(IQRate)
 			amp := 0.42
 			re := amp * math.Cos(phaseMain)
 			im := amp * math.Sin(phaseMain)
 
 			// Two drifting carriers ±60-90 kHz off center.
-			fA := 65e3 + 8e3*math.Sin(2*math.Pi*0.05*float64(t)/IQRate)
-			fB := -82e3 + 6e3*math.Sin(2*math.Pi*0.03*float64(t)/IQRate)
-			phaseA += 2 * math.Pi * fA / IQRate
-			phaseB += 2 * math.Pi * fB / IQRate
+			fA := 65e3 + 8e3*math.Sin(2*math.Pi*0.05*float64(t)/float64(IQRate))
+			fB := -82e3 + 6e3*math.Sin(2*math.Pi*0.03*float64(t)/float64(IQRate))
+			phaseA += 2 * math.Pi * fA / float64(IQRate)
+			phaseB += 2 * math.Pi * fB / float64(IQRate)
 			re += 0.20*math.Cos(phaseA) + 0.16*math.Cos(phaseB)
 			im += 0.20*math.Sin(phaseA) + 0.16*math.Sin(phaseB)
